style: move errors to more appropriate place and fix import cicle not alowed error
