+++ pkg/domain/accounts/usecases/get_all_test.go
@@ -1,4 +1,4 @@
-package accounts
+package usecases
 
 import (
 	"testing"
