+++ pkg/domain/accounts/usecases/get_balance_by_id.go
@@ -1,4 +1,4 @@
-package accounts
+package usecases
 
 func (au AccountUseCase) GetBalanceByID(id string) (int, error) {
 	return au.storage.GetBalanceByID(id)
