+++ pkg/domain/accounts/usecases/get_balance_by_id_test.go
@@ -1,4 +1,4 @@
-package accounts
+package usecases
 
 import (
 	"testing"
