+++ pkg/domain/accounts/usecases/new_use_case.go
@@ -1,0 +1,15 @@
+package usecases
+
+import (
+	accounts_usecase "github.com/daniel1sender/Desafio-API/pkg/domain/accounts"
+)
+
+type AccountUseCase struct {
+	storage accounts_usecase.Repository
+}
+
+func NewUseCase(storage accounts_usecase.Repository) AccountUseCase {
+	return AccountUseCase{
+		storage: storage,
+	}
+}
