+++ pkg/domain/accounts/usecases/update_balance.go
@@ -1,4 +1,4 @@
-package accounts
+package usecases
 
 import (
 	"errors"
