+++ pkg/gateways/http/accounts/get_balance_by_id_test.go
@@ -6,9 +6,9 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/daniel1sender/Desafio-API/pkg/domain/accounts"
 	accounts_usecase "github.com/daniel1sender/Desafio-API/pkg/domain/accounts"
 	server_http "github.com/daniel1sender/Desafio-API/pkg/gateways/http"
-	"github.com/daniel1sender/Desafio-API/pkg/gateways/store"
 )
 
 func TestGetBalanceByID(t *testing.T) {
@@ -44,7 +44,7 @@ func TestGetBalanceByID(t *testing.T) {
 
 		expectedBalance := 0
 		useCase := accounts_usecase.UseCaseMock{Balance: expectedBalance,
-			Error: store.ErrIDNotFound}
+			Error: accounts.ErrIDNotFound}
 		h := NewHandler(&useCase)
 
 		newRequest, _ := http.NewRequest(http.MethodGet, "accounts/{id}/balance", nil)
@@ -63,8 +63,8 @@ func TestGetBalanceByID(t *testing.T) {
 			t.Errorf("expected '%s' but got '%s'", server_http.JSONContentType, newResponse.Header().Get("content-type"))
 		}
 
-		if response.Reason != store.ErrIDNotFound.Error() {
-			t.Errorf("expected '%s' but got '%s'", store.ErrIDNotFound.Error(), response.Reason)
+		if response.Reason != accounts.ErrIDNotFound.Error() {
+			t.Errorf("expected '%s' but got '%s'", accounts.ErrIDNotFound.Error(), response.Reason)
 		}
 
 	})
