+++ pkg/gateways/store/memory/accounts/get_cpf.go
@@ -1,11 +1,13 @@
 package accounts
 
-import "github.com/daniel1sender/Desafio-API/pkg/gateways/store"
+import (
+	"github.com/daniel1sender/Desafio-API/pkg/domain/accounts"
+)
 
 func (s AccountStorage) CheckCPF(cpf string) error {
 	for _, storedAccount := range s.storage {
 		if storedAccount.CPF == cpf {
-			return store.ErrExistingCPF
+			return accounts.ErrExistingCPF
 		}
 	}
 	return nil
