+++ pkg/gateways/store/repository/accounts/check_cpf.go
@@ -1,11 +1,13 @@
 package accounts
 
-import "github.com/daniel1sender/Desafio-API/pkg/gateways/store"
+import (
+	"github.com/daniel1sender/Desafio-API/pkg/domain/accounts"
+)
 
 func (ar accountRepository) CheckCPF(cpf string) error {
 	for _, value := range ar.users {
 		if value.CPF == cpf {
-			return store.ErrExistingCPF
+			return accounts.ErrExistingCPF
 		}
 	}
 	return nil
